requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
Pillow>=10.2.0
python-dotenv>=1.0.1
streamlit-authenticator>=0.3.3
reportlab>=4.0.0
requests>=2.31.0
bcrypt>=4.1.2
groq>=0.5.0
pyngrok>=7.1.0
supabase>=2.4.0